pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "routes>=2.5.1",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
    "oauthlib>=3.2.2",
    "pyjwt>=2.10.1",
    "psutil>=7.0.0",
    "requests>=2.32.3",
    "schedule>=1.2.2",
]
